+++ internal/model/user.go
@@ -19,7 +19,7 @@ type User struct {
 }
 
 type UserInput struct {
-	ID           int32     `json:"id" db:"id"`
+	ID           int64     `json:"id" db:"id"`
 	FirstName    string    `json:"firstName" db:"firstName"`
 	LastName     string    `json:"lastName" db:"lastName"`
 	Email        string    `json:"email" db:"email"`
@@ -31,13 +31,13 @@ type UserResponse struct {
 	FirstName string          `json:"firstName" db:"firstName"`
 	LastName  string          `json:"lastName" db:"lastName"`
 	Email     string          `json:"email" db:"email"`
-	ID        int32           `json:"id" db:"id"`
+	ID        int64           `json:"id" db:"id"`
 	Balance   decimal.Decimal `json:"balance" db:"balance"`
 	CreatedAt time.Time       `json:"createdAt" db:"createdAt"`
 	DeletedAt time.Time       `json:"deletedAt" db:"deletedAt"`
 }
 
 type Balance struct {
-	ID      int32
+	ID      int64
 	Amount decimal.Decimal
 }
